# Perf/reduce memory allocs