--- a/utils.go
+++ b/utils.go
@@ -15,8 +15,8 @@ const (
 var matchFirstCap = regexp.MustCompile("(.)([A-Z][a-z]+)")
 var matchAllCap = regexp.MustCompile("([a-z0-9])([A-Z])")
 
-func convertToSnakeCase(name string) string {
-	snake := matchFirstCap.ReplaceAllString(name, "${1}_${2}")
+func convertToSnakeCase(name string) (snake string) {
+	snake = matchFirstCap.ReplaceAllString(name, "${1}_${2}")
 	snake = matchAllCap.ReplaceAllString(snake, "${1}_${2}")
 	return strings.ToLower(snake)
 }