- reduce memory allocation
- add utils test

**before**
```
goos: darwin
goarch: amd64
pkg: github.com/teerapon19/go-query-string
cpu: VirtualApple @ 2.50GHz
BenchmarkConvertToSnakeCase-8            4077366               284.7 ns/op           171 B/op          9 allocs/op
BenchmarkUnmarshal-8                     1200618               988.2 ns/op          1035 B/op         43 allocs/op
BenchmarkMarshal-8                       1458153               817.4 ns/op           829 B/op         40 allocs/op
PASS
ok      github.com/teerapon19/go-query-string   6.372s
```

**after**
```
goos: darwin
goarch: amd64
pkg: github.com/teerapon19/go-query-string
cpu: VirtualApple @ 2.50GHz
BenchmarkConvertToSnakeCase-8            4151071               283.3 ns/op           171 B/op          9 allocs/op
BenchmarkUnmarshal-8                     1203369               977.8 ns/op          1033 B/op         43 allocs/op
BenchmarkMarshal-8                       1643745               711.0 ns/op           561 B/op         36 allocs/op
PASS
ok      github.com/teerapon19/go-query-string   6.377s
```