@teerapon19 2023-11-22
Thank you for helping approving reducing the alloc memory. It's improving 10% for BenchmarkMarshal-8.